requirements/prod.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/prod.in
#
absl-py==0.12.0
    # via tensorboard
aiohttp==3.7.4.post0
    # via fsspec
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
boltons==20.2.1
    # via -r requirements/prod.in
cached-property==1.5.2
    # via h5py
cachetools==4.2.1
    # via google-auth
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via flask
dataclasses==0.8
    # via torch
editdistance==0.5.3
    # via -r requirements/prod.in
flask==1.1.2
    # via -r requirements/prod.in
fsspec[http]==0.8.7
    # via pytorch-lightning
future==0.18.2
    # via pytorch-lightning
google-auth-oauthlib==0.4.4
    # via tensorboard
google-auth==1.28.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
grpcio==1.36.1
    # via tensorboard
h5py==3.1.0
    # via -r requirements/prod.in
idna-ssl==1.1.0
    # via aiohttp
idna==2.10
    # via
    #   idna-ssl
    #   requests
    #   yarl
importlib-metadata==3.10.0
    # via
    #   fsspec
    #   markdown
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
markdown==3.3.4
    # via tensorboard
markupsafe==1.1.1
    # via jinja2
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
numpy==1.19.5
    # via
    #   -r requirements/prod.in
    #   h5py
    #   pytorch-lightning
    #   tensorboard
    #   torch
    #   torchvision
oauthlib==3.1.0
    # via requests-oauthlib
pillow==8.2.0
    # via torchvision
protobuf==3.15.7
    # via tensorboard
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pytorch-lightning==1.2.6
    # via -r requirements/prod.in
pyyaml==5.3.1
    # via pytorch-lightning
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   -r requirements/prod.in
    #   fsspec
    #   requests-oauthlib
    #   tensorboard
rsa==4.7.2
    # via google-auth
six==1.15.0
    # via
    #   absl-py
    #   google-auth
    #   grpcio
    #   protobuf
    #   tensorboard
smart-open==5.0.0
    # via -r requirements/prod.in
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard==2.4.1
    # via pytorch-lightning
toml==0.10.2
    # via -r requirements/prod.in
torch==1.7.1
    # via
    #   -r requirements/prod.in
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==0.2.0
    # via pytorch-lightning
torchvision==0.8.2
    # via -r requirements/prod.in
tqdm==4.59.0
    # via
    #   -r requirements/prod.in
    #   pytorch-lightning
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   importlib-metadata
    #   torch
    #   yarl
urllib3==1.26.4
    # via requests
werkzeug==1.0.1
    # via
    #   flask
    #   tensorboard
wheel==0.36.2
    # via tensorboard
yarl==1.6.3
    # via aiohttp
zipp==3.4.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
